chore: remove root /supabase (stale functions/migrations/policies); retarget SDK to shared/supabase; move authHelpers into storefronts
