+++ storefronts/tests/adapters/checkout.test.js
@@ -12,7 +12,7 @@ vi.mock('@supabase/supabase-js', () => {
   const from = vi.fn(() => ({ select }));
   return { createClient: vi.fn(() => ({ from })) };
 });
-vi.mock('../../../supabase/browserClient.js', () => {
+vi.mock('../../../shared/supabase/browserClient.js', () => {
   const eq = vi.fn(() => ({ eq, maybeSingle: supabaseMaybeSingle }));
   const select = vi.fn(() => ({ eq }));
   const from = vi.fn(() => ({ select }));
