+++ storefronts/tests/sdk/cart-dom-trigger.test.js
@@ -15,7 +15,7 @@ describe("cart DOM trigger", () => {
       Promise.resolve({ ok: true, json: () => Promise.resolve({}) })
     );
     global.console = { log: vi.fn(), warn: vi.fn() };
-    vi.doMock("../../supabase/browserClient.js", () => ({
+    vi.doMock("../../shared/supabase/browserClient.js", () => ({
       supabase: {
         from: vi.fn(() => ({
           select: vi.fn(() => ({
@@ -32,7 +32,7 @@ describe("cart DOM trigger", () => {
   });
 
   afterEach(() => {
-    vi.doUnmock("../../supabase/browserClient.js");
+    vi.doUnmock("../../shared/supabase/browserClient.js");
     vi.doUnmock("../../features/auth/init.js");
     vi.doUnmock("../../features/currency/index.js");
     vi.doUnmock("../../features/cart/init.js");
