--- supabase/authHelpers.js
@@ -1,2 +1,0 @@
-export * from './client/authHelpers.js';
-
