--- supabase/browserClient.js
@@ -1,3 +1,0 @@
-export * from './client/browserClient.js';
-export { default } from './client/browserClient.js';
-
