--- supabase/config.toml
@@ -1,22 +1,0 @@
-[functions.proxy-live-rates]
-verify_jwt = false
-
-[functions.webflow-order-handler]
-verify_jwt = false
-
-[functions.get_gateway_credentials]
-enabled = true
-verify_jwt = true
-import_map = "./functions/get_gateway_credentials/deno.json"
-# Uncomment to specify a custom file path to the entrypoint.
-# Supported file extensions are: .ts, .js, .mjs, .jsx, .tsx
-entrypoint = "./functions/get_gateway_credentials/index.ts"
-# Specifies static files to be bundled with the function. Supports glob patterns.
-# For example, if you want to serve static HTML pages in your function:
-# static_files = [ "./functions/get_gateway_credentials/*.html" ]
-
-[functions.get_public_store_settings]
-enabled = true
-verify_jwt = true
-import_map = "./functions/get_public_store_settings/deno.json"
-entrypoint = "./functions/get_public_store_settings/index.ts"
