--- supabase/functions/.env.example
@@ -1,8 +1,0 @@
-# Supabase edge function environment variables
-# Token from https://openexchangerates.org/
-OPENEXCHANGERATES_TOKEN=
-# Required token: YOUR_TOKEN_HERE
-
-# Supabase credentials for webflow-order-handler
-SUPABASE_URL=
-SUPABASE_SERVICE_ROLE_KEY=
