--- supabase/functions/dummy.test.ts
@@ -1,7 +1,0 @@
-import { describe, it, expect } from 'vitest';
-
-describe('dummy', () => {
-  it('placeholder', () => {
-    expect(true).toBe(true);
-  });
-});
