--- supabase/functions/get_gateway_credentials/.npmrc
@@ -1,3 +1,0 @@
-# Configuration for private npm package dependencies
-# For more information on using private registries with Edge Functions, see:
-# https://supabase.com/docs/guides/functions/import-maps#importing-from-private-registries
