--- supabase/functions/get_gateway_credentials/deno.json
@@ -1,3 +1,0 @@
-{
-  "imports": {}
-}
