--- supabase/functions/get_public_store_settings/deno.json
@@ -1,3 +1,0 @@
-{
-  "imports": {}
-}
