--- supabase/functions/proxy-live-rates/deno.json
@@ -1,1 +1,0 @@
-{}
