--- supabase/functions/proxy-live-rates/index.ts
@@ -1,3 +1,0 @@
-import { handleRequest } from "./handler.ts";
-
-Deno.serve(handleRequest);
