--- supabase/functions/webflow-order-handler/deno.json
@@ -1,1 +1,0 @@
-{}
